Odd or Even project
Project Description:
This project serves as a foundation for learning and improving my basic C programming skills. The main functionality of the project is to implement a program that separates numbers into even and odd categories. It demonstrates the use of conditional statements, loops, and basic input/output operations in C.